requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-JWT-Extended==4.5.2
psycopg2-binary==2.9.11
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
